app restructuring, removed unnecessary code
