+++ src/components/Result/Image/Image.module.css
@@ -1,6 +1,5 @@
 .ImageBox {
    margin: 10px 5px 0 10px;
-   border-right: 2px solid white;
 }
 
 .Image {
